+++ src/Main.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
 
-        int myIndex = (int) service.calculate(77.54, 1.98);
+        int myIndex = service.calculate(77.54, 1.98);
 
         System.out.println(myIndex);
     }
